Move misc into utils (#2688) Bruception
