+++ frontend/src/scripts/config-validation.ts
@@ -1,4 +1,4 @@
-import * as Misc from "./misc";
+import * as Misc from "./utils/misc";
 import * as Notifications from "./elements/notifications";
 
 type PossibleType =
