+++ frontend/src/scripts/controllers/chart-controller.js
@@ -1,7 +1,7 @@
 import Chart from "chart.js";
 import * as TestInput from "../test/test-input";
 import * as ThemeColors from "../elements/theme-colors";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import Config from "../config";
 import * as ConfigEvent from "../observables/config-event";
