+++ frontend/src/scripts/elements/account-button.ts
@@ -1,4 +1,4 @@
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 
 export function loading(truefalse: boolean): void {
   if (truefalse) {
