+++ frontend/src/scripts/elements/leaderboards.ts
@@ -1,7 +1,7 @@
 import Ape from "../ape";
 import * as DB from "../db";
 import Config from "../config";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import * as Notifications from "./notifications";
 
 const currentLeaderboard = "time_15";
