+++ frontend/src/scripts/elements/notifications.ts
@@ -1,4 +1,4 @@
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 
 let id = 0;
 class Notification {
