+++ frontend/src/scripts/pages/account.ts
@@ -12,7 +12,7 @@ import * as SignOutButton from "../account/sign-out-button";
 import * as TodayTracker from "../test/today-tracker";
 import * as Notifications from "../elements/notifications";
 import Page from "./page";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import * as ActivePage from "../states/active-page";
 
 let filterDebug = false;
