+++ frontend/src/scripts/popups/quote-search-popup.ts
@@ -6,7 +6,7 @@ import * as Notifications from "../elements/notifications";
 import * as QuoteSubmitPopup from "./quote-submit-popup";
 import * as QuoteApprovePopup from "./quote-approve-popup";
 import * as QuoteReportPopup from "./quote-report-popup";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import {
   buildSearchService,
   SearchService,
