+++ frontend/src/scripts/ready.ts
@@ -1,6 +1,6 @@
 import * as ManualRestart from "./test/manual-restart-tracker";
 import Config, * as UpdateConfig from "./config";
-import * as Misc from "./misc";
+import * as Misc from "./utils/misc";
 import * as VerificationController from "./controllers/verification-controller";
 import * as RouteController from "./controllers/route-controller";
 import * as PageController from "./controllers/page-controller";
