+++ frontend/src/scripts/test/layout-emulator.ts
@@ -1,5 +1,5 @@
 import Config from "../config";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import { capsState } from "./caps-warning";
 
 export async function getCharFromEvent(
