+++ frontend/src/scripts/test/pace-caret.ts
@@ -4,7 +4,7 @@ import * as TestUI from "./test-ui";
 import Config from "../config";
 import * as DB from "../db";
 import * as SlowTimer from "../states/slow-timer";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import * as TestActive from "../states/test-active";
 import * as TestState from "./test-state";
 import * as ConfigEvent from "../observables/config-event";
