+++ frontend/src/scripts/test/test-ui.ts
@@ -8,7 +8,7 @@ import * as CustomText from "./custom-text";
 import * as Caret from "./caret";
 import * as OutOfFocus from "./out-of-focus";
 import * as Replay from "./replay";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import * as SlowTimer from "../states/slow-timer";
 import * as ConfigEvent from "../observables/config-event";
