+++ frontend/src/scripts/test/timer-progress.ts
@@ -1,6 +1,6 @@
 import Config from "../config";
 import * as CustomText from "./custom-text";
-import * as Misc from "../misc";
+import * as Misc from "../utils/misc";
 import * as TestWords from "./test-words";
 import * as TestInput from "./test-input";
 import * as Time from "../states/time";
